move files to clarify structure
