+++ src/databricks/labs/remorph/config.py
@@ -5,7 +5,10 @@ from sqlglot.dialects.dialect import Dialect, Dialects, DialectType
 
 from databricks.labs.remorph.helpers.morph_status import ParserError
 from databricks.labs.remorph.reconcile.recon_config import Table
-from databricks.labs.remorph.snow import databricks, oracle, snowflake, presto
+from databricks.labs.remorph.transpiler.generator import databricks
+from databricks.labs.remorph.transpiler.snowflake import snowflake
+from databricks.labs.remorph.transpiler.oracle import oracle
+from databricks.labs.remorph.transpiler.presto import presto
 
 logger = logging.getLogger(__name__)
