+++ src/databricks/labs/remorph/transpiler/sqlglot/lca_utils.py
@@ -9,7 +9,7 @@ from sqlglot.expressions import Expression, Select
 from sqlglot.optimizer.scope import Scope, build_scope
 
 from databricks.labs.remorph.helpers.morph_status import ValidationError
-from databricks.labs.remorph.snow.local_expression import AliasInfo
+from databricks.labs.remorph.transpiler.sqlglot.local_expression import AliasInfo
 
 logger = logging.getLogger(__name__)
